// src/BLL/models/trees-requests-create/trees-requests.types.ts
export type TreesRequestsCreateFormValues = {
    name: Undefinedtable<string>;
    description: Undefinedtable<string>;
    coordinates: {
        accuracy: Undefinedtable<number>;
        latitude: Undefinedtable<number>;
        longitude: Undefinedtable<number>;
    };
    image: Undefinedtable<File>;
};